# Conflicts:
#	ethcore/light/src/types/mod.rs
#	js/package.json
#	js/src/i18n/en/index.js
#	js/src/modals/SMSVerification/GatherData/gatherData.js
#	js/src/modals/SMSVerification/SendConfirmation/sendConfirmation.js
#	js/src/modals/SMSVerification/SendRequest/sendRequest.js
#	js/src/ui/index.js
#	js/src/util/nullable-proptype.js
#	js/src/util/proptypes.js
#	js/src/views/Settings/Parity/parity.js

+++ sync/src/chain.rs
@@ -192,7 +192,7 @@ pub struct SyncStatus {
 	/// Syncing protocol version. That's the maximum protocol version we connect to.
 	pub protocol_version: u8,
 	/// The underlying p2p network version.
-	pub network_id: usize,
+	pub network_id: u64,
 	/// `BlockChain` height for the moment the sync started.
 	pub start_block_number: BlockNumber,
 	/// Last fully downloaded and imported block number (if any).
@@ -273,7 +273,7 @@ struct PeerInfo {
 	/// Peer chain genesis hash
 	genesis: H256,
 	/// Peer network id
-	network_id: usize,
+	network_id: u64,
 	/// Peer best block hash
 	latest_hash: H256,
 	/// Peer total difficulty if known
@@ -341,7 +341,7 @@ pub struct ChainSync {
 	/// Last propagated block number
 	last_sent_block_number: BlockNumber,
 	/// Network ID
-	network_id: usize,
+	network_id: u64,
 	/// Optional fork block to check
 	fork_block: Option<(BlockNumber, H256)>,
 	/// Snapshot downloader.
